+++ Project/src/main/java/com/example/backestopenda/controllers/StudentController.java
@@ -2,6 +2,8 @@ package com.example.backestopenda.controllers;
 
 import com.example.backestopenda.dto.StudentDto;
 import com.example.backestopenda.services.StudentService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,18 +23,21 @@ import java.net.URI;
 
 @RestController
 @RequestMapping("/student")
+@Tag(name = "Student")
 public class StudentController {
 
     @Autowired
     private StudentService studentService;
 
     @GetMapping
+    @Operation(summary = "Finds all students")
     public ResponseEntity<Page<StudentDto>> findAll(@PageableDefault() Pageable pageable) {
 
         return ResponseEntity.ok(studentService.findAll(pageable));
     }
 
     @GetMapping("/{id}")
+    @Operation(summary = "Finds an student by id")
     public ResponseEntity<StudentDto> findById(@PathVariable(value = "id") Long id) {
         StudentDto studentDto = studentService.findById(id);
         if (studentDto == null){
@@ -42,6 +47,7 @@ public class StudentController {
     }
 
     @PostMapping
+    @Operation(summary = "Creates an student")
     public ResponseEntity<StudentDto> create(@RequestBody StudentDto dto, UriComponentsBuilder uriBuilder){
         StudentDto studentDto = studentService.create(dto);
         URI path = uriBuilder.path("/student/{id}").buildAndExpand(studentDto.getStudentId()).toUri();
@@ -50,14 +56,19 @@ public class StudentController {
     }
 
     @PutMapping("/{id}")
+    @Operation(summary = "Updates an student")
     public ResponseEntity<StudentDto> update(@PathVariable Long id, @RequestBody StudentDto dto){
         StudentDto studentDto = studentService.update(id, dto);
         return ResponseEntity.ok(studentDto);
     }
 
     @DeleteMapping("/{id}")
+    @Operation(summary = "Deletes an student")
     public ResponseEntity<?> delete(@PathVariable Long id){
-        studentService.delete(id);
-        return ResponseEntity.noContent().build();
+        if (studentService.existsById(id)){
+            studentService.delete(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
